add load_input module for reading input files
