+++ barrier3d/__init__.py
@@ -1,5 +1,6 @@
 from .barrier3d import Barrier3d
-from .configuration import Barrier3dConfiguration, load_inputs
+from .configuration import Barrier3dConfiguration
+from .load_input import load_inputs
 
 
-__all__ = ["Barrier3d", "load_inputs"]
+__all__ = ["Barrier3d", "Barrier3dConfiguration", "load_inputs"]
