+++ barrier3d/configuration.py
@@ -1,3 +1,5 @@
+import contextlib
+import os
 import pathlib
 
 import numpy as np
@@ -25,7 +27,7 @@ class Barrier3dConfiguration(Configuration):
     )
     BarrierLength = FloatField(
         "BarrierLength",
-        default=1000.0,
+        default=300.0,
         units="m",
         description="Static length (alongshore) of island segment",
         validators=[Range(lower=0)],
@@ -53,7 +55,7 @@ class Barrier3dConfiguration(Configuration):
     )
     BayDepth = FloatField(
         "BayDepth",
-        default=1.5,
+        default=2.0,
         units="m",
         description="Depth of bay benind island segment",
         validators=[Range(lower=0)],
@@ -81,24 +83,24 @@ class Barrier3dConfiguration(Configuration):
     )
     rmin = FloatField(
         "rmin",
-        default=0.35,
+        default=0.05,
         description="Minimum growth rate for logistic dune growth",
     )
     rmax = FloatField(
         "rmax",
-        default=0.85,
+        default=0.55,
         description="Maximum growth rate for logistic dune growth",
     )
     HdDiffu = FloatField(
         "HdDiffu",
-        default=0.8,
+        default=0.45,
         units="m",
         description="Dune diffusion parameter (i.e. max height offset between adjacent dune cells)",
         validators=[Range(lower=0)],
     )
     Dmaxel = FloatField(
         "Dmaxel",
-        default=3.0,
+        default=2.3,
         units="m",
         description="Maximum elevation of dunes",
         validators=[Range(lower=0)],
@@ -124,7 +126,7 @@ class Barrier3dConfiguration(Configuration):
     )
     Rat = FloatField(
         "Rat",
-        default=-1,
+        default=0,
         units="m / y",
         description="Rate of shoreline reatreat attributed to alongshore transport; (-) = erosion, (+) = accretion",
     )
@@ -154,54 +156,54 @@ class Barrier3dConfiguration(Configuration):
     )
     surge_tide_m = FloatField(
         "surge_tide_m",
-        default=0.91,
+        default=0.0,
         units="m",
         description="Mean storm surge + tide water levels",
         validators=[Range(lower=0)],
     )
     surge_tide_sd = FloatField(
         "surge_tide_sd",
-        default=0.3,
+        default=0.0,
         units="m",
         description="Standard deviation of surge + tide water levels",
         validators=[Range(lower=0)],
     )
     height_mu = FloatField(
         "height_mu",
-        default=1.02,
+        default=0.0,
         units="m",
         description="Mu storm wave height (lognormal distribution)",
         validators=[Range(lower=0)],
     )
     height_sigma = FloatField(
         "height_sigma",
-        default=0.19,
+        default=0.0,
         units="m",
         description="Sigma storm wave height (lognormal distribution)",
         validators=[Range(lower=0)],
     )
     period_m = FloatField(
         "period_m",
-        default=10.74,
+        default=0.0,
         description="Mean storm wave period",
         validators=[Range(lower=0)],
     )
     period_sd = FloatField(
         "period_sd",
-        default=2.76,
+        default=0.0,
         description="Standard deviation of storm wave periods",
         validators=[Range(lower=0)],
     )
     duration_mu = FloatField(
         "duration_mu",
-        default=2.9249,
+        default=0.0,
         units="hr",
         description="Mu storm duration (lognormal distribution)",
         validators=[Range(lower=0)],
     )
     duration_sigma = FloatField(
         "duration_sigma",
-        default=0.638673,
+        default=0.0,
         units="hr",
         description="Sigma storm duration (lognormal distribution)",
         validators=[Range(lower=0)],
@@ -214,7 +216,7 @@ class Barrier3dConfiguration(Configuration):
     )
     StormTimeSeries = BooleanField(
         "StormTimeSeries",
-        default=0,
+        default=True,
         description="Storms will come from a time series",
     )
     StormSeries = ArrayField(
@@ -234,17 +236,17 @@ class Barrier3dConfiguration(Configuration):
     )
     Rin_r = FloatField(
         "Rin_r",
-        default=6,
+        default=2,
         description="Run-up regime infiltration rate (volume of overwash flow lost per m cross-shore per time step)",
     )
     Rin_i = FloatField(
         "Rin_i",
-        default=1,
+        default=0.25,
         description="Inundation regime infiltration rate (volume of overwash flow lost per m cross-shore per time step)",
     )
     Qs_min = FloatField(
         "Qs_min",
-        default=0.05,
+        default=1.0,
         units="m^3 / hr",
         description="Minimum discharge needed for sediment transport",
     )
@@ -256,24 +258,24 @@ class Barrier3dConfiguration(Configuration):
     )
     Kr = FloatField(
         "Kr",
-        default=0.0001,
+        default=0.0003,
         description="Sediment flux constant, run-up regime",
     )
     Ki = FloatField(
         "Ki",
-        default=0.0004,
+        default=0.000075,
         description="Sediment flux constant, inundation regime",
     )
     Cbb_r = FloatField(
         "Cbb_r",
-        default=0.65,
+        default=0.5,
         units="",
         description="Coefficient for exponential decay of sediment load entering back-barrier bay in run-up regime",
         validators=[Range(lower=0, upper=1)],
     )
     Cbb_i = FloatField(
         "Cbb_i",
-        default=0.9,
+        default=0.8,
         units="",
         description="Coefficient for exponential decay of sediment load entering back-barrier bay in inundation regime",
         validators=[Range(lower=0, upper=1)],
@@ -375,7 +377,7 @@ class Barrier3dConfiguration(Configuration):
     )
     ShrubEl_min = FloatField(
         "ShrubEl_min",
-        default=1.2,
+        default=0.6,
         units="m",
         description="Elevation range for shrub growth, minimum bound",
         validators=[Range(lower=0)],
@@ -427,145 +429,6 @@ class Barrier3dConfiguration(Configuration):
             )
         )
 
-    @staticmethod
-    def from_xlsx(path_to_xlsx):
-        from .parameters import from_xlsx
-
-        return from_xlsx(path_to_xlsx)
-
-
-def load_inputs(path_to_folder, prefix="barrier3d", fmt="yaml"):
-    """Load input parameter and elevation data for Barrier3d."""
-    base_dir = pathlib.Path(path_to_folder)
-    if fmt == "yaml":
-        path_to_parameters = f"{prefix}-parameters.yaml"
-        path_to_elevations = f"{prefix}-elevations.csv"
-        # path_to_parameters = "parameters.yaml"
-        # path_to_elevations = "elevations.csv"
-    elif fmt == "xlsx":
-        path_to_parameters = f"{prefix}.xlsx"
-        path_to_elevations = f"{prefix}.xlsx"
-    elif fmt == "py":
-        path_to_parameters = f"{prefix}-parameters.py"
-        path_to_elevations = f"{prefix}-elevations.csv"
-    else:
-        raise ValueError(
-            "unrecognized format for input files ({fmt} not in {fmts})".format(
-                fmt=fmt, fmts=", ".join(fmts)
-            )
-
-        )
-
-    if base_dir.is_file():
-        path_to_parameters = base_dir.name
-        base_dir = base_dir.parent
-
-    params = load_configuration(base_dir / path_to_parameters, fmt=fmt)
-    params["InteriorDomain"] = load_elevation(
-        base_dir / path_to_elevations, fmt="csv" if fmt != "xlsx" else "xlsx"
-    )
-
-    return _process(params)
-
-
-def load_configuration(path_to_file, fmt="yaml"):
-    if fmt == "yaml":
-        return Barrier3dConfiguration.from_yaml(path_to_file).data
-    elif fmt == "py":
-        return Barrier3dConfiguration.from_py(path_to_file).data
-    else:
-        raise ValueError(f"format not understood ({fmt})")
-    # return Barrier3dConfiguration.from_path(path_to_file, fmt=fmt).data
-
-
-def load_storms(path_to_file, fmt="csv"):
-    if fmt == "npy":
-        data = np.load(path_to_file)
-    elif fmt == "csv":
-        df = pandas.read_csv(
-            path_to_file,
-            names=("time", "Rhigh", "Rlow", "period", "duration"),
-            dtype={
-                "time": int,
-                "Rhigh": float,
-                "Rlow": float,
-                "period": float,
-                "duration": int,
-            },
-            comment="#",
-            header=0,
-        )
-        data = np.hstack(
-            (
-                df["time"][:, None],
-                df["Rhigh"][:, None],
-                df["Rlow"][:, None],
-                df["period"][:, None],
-                df["duration"][:, None],
-            )
-        )
-    else:
-        fmts = ", ".join(["npy", "csv"])
-        raise ValueError(
-            f"unrecognized format for storm time series ({fmt} not one of {fmts})"
-        )
-
-    return data
-
-
-def load_dunes(path_to_file, fmt="csv"):
-    if fmt == "npy":
-        data = np.load(path_to_file)
-    elif fmt == "csv":
-        data = pandas.read_csv(
-            path_to_file, names=("DuneStart",), dtype=float, comment="#", header=0
-        )["DuneStart"].values
-    else:
-        fmts = ", ".join(["npy", "csv"])
-        raise ValueError(
-            f"unrecognized format for dunes ({fmt} not one of {fmts})"
-        )
-
-    return data
-
-
-def load_growth_param(path_to_file, fmt="csv"):
-    if fmt == "npy":
-        data = np.load(path_to_file)
-    elif fmt == "csv":
-        data = pandas.read_csv(
-            path_to_file, names=("growth_param",), dtype=float, comment="#", header=0
-        )["growth_param"].values
-    else:
-        fmts = ", ".join(["npy", "csv"])
-        raise ValueError(
-            f"unrecognized format for growth param ({fmt} not one of {fmts})"
-        )
-
-    return data
-
-
-def load_elevation(path_to_file, fmt="csv"):
-    loaders = {
-        "csv": _load_initial_elevation_csv,
-        "xlsx": _load_initial_elevation_xlsx,
-    }
-    try:
-        loader = loaders[fmt]
-    except KeyError:
-        fmts = ", ".join(loaders.keys())
-        raise ValueError(f"unrecognized format for elevation ({fmt} not one of {fmts})")
-
-    def _process(data):
-        shape = data.x.max() + 1, data.y.max() + 1
-        elevation = np.zeros(shape, dtype=float)
-        elevation[(data.x, data.y)] = data.z
-
-        return elevation
-        # return np.flipud(np.rot90(elevation))
-
-    return _process(loader(path_to_file))
-
 
 def _load_initial_elevation_xlsx(path_to_xlsx):
     wb = xlrd.open_workbook(path_to_xlsx)
@@ -588,127 +451,3 @@ def _load_initial_elevation_csv(path_to_csv):
         comment="#",
         header=0,
     )
-
-
-def _process(params):
-    params["TMAX"] = int(params["TMAX"]) + 1
-    params["StormStart"] = int(params["StormStart"])
-
-    if params["StormTimeSeries"]:
-        params["StormSeries"] = load_storms("barrier3d-storms.csv")
-    else:
-        params["StormSeries"] = []
-
-    params["LShoreface"] /= 10.0
-    params["DShoreface"] /= 10.0
-    params["BayDepth"] /= 10.0
-    params["MHW"] /= 10.0
-
-    # intElev = load_initial_elevation(morph)
-
-    params["BarrierLength"] = int(params["BarrierLength"] / 10.0)
-    params["DuneWidth"] = int(params["DuneWidth"] / 10.0)
-
-    z = params["InteriorDomain"] / 10.0 - params["MHW"]
-    z[z <= 0.0] = -params["BayDepth"]
-    # above_water = np.any(z > 0, axis=0)
-    # z = z[:, above_water]
-    # NOTE: if we haven't already flipped/rotated
-    u = 1
-    while u == 1:  # Remove all rows that have zero subaerial cells
-        if all(z[:, -1] <= 0):
-            z = z[:, :-1]
-        else:
-            u = 0
-    # while np.all(z[:, -1] <= 0):
-    #     z = z[:, :-1]
-    # NOTE: if we have already flipped/rotated
-    # u = 1
-    # while u == 1:  # Remove all rows that have zero subaerial cells
-    #     if all(z[0, :] <= 0):
-    #         z = z[1:, :]
-    #     else:
-    #         u = 0
-
-    # NOTE: We do this in the process step
-    params["InteriorDomain"] = np.flipud(np.rot90(z))
-
-    if len(params["InteriorDomain"][0]) > params["BarrierLength"]:
-        # Reduce to specified max length
-        params["InteriorDomain"] = params["InteriorDomain"][
-            :, : params["BarrierLength"]
-        ]
-    else:
-        params["BarrierLength"] = len(params["InteriorDomain"][0])
-
-    # intElev = intElev / 10.0 - params["MHW"]
-    # intElev[intElev <= 0.0] = - params["BayDepth"]
-
-    # NOTE: This removes ALL columns that have zero subaerial cells
-    # intElev = intElev[:, np.all(intElev <= 0, axis=0)]
-
-    # Flip to correct orientation
-    # params["InteriorDomain"] = np.flipud(np.rot90(intElev))
-
-    params["DomainWidth"] = len(params["InteriorDomain"])
-
-    params["Dstart"] /= 10.0
-    params["BermEl"] = params["BermEl"] / 10.0 - params["MHW"]
-
-    # Initialize dune crest height domain
-    params["DuneDomain"] = np.zeros(
-        (params["TMAX"], params["BarrierLength"], params["DuneWidth"])
-    )
-    if params["StormTimeSeries"]:
-        dune_start = load_dunes("barrier3d-dunes.csv")
-        params["DuneDomain"][0, :, 0] = dune_start[0:params["BarrierLength"]]    
-    else:
-        params["DuneDomain"][0, :, 0] = np.ones([1, params["BarrierLength"]]) * (
-            params["Dstart"]
-            + (-0.01 + (0.01 - (-0.01)) * np.random.rand(1, params["BarrierLength"]))
-        )
-    params["DuneDomain"][0, :, 1:] = params["DuneDomain"][0, :, 0, None]
-
-    if params["StormTimeSeries"]:
-        growth_param_start = load_growth_param("barrier3d-growthparam.csv")
-        params["growthparam"] = growth_param_start[0: params["BarrierLength"]]
-    else:
-        params["growthparam"] = params["rmin"] + (
-            params["rmax"] - params["rmin"]
-        ) * np.random.rand(1, params["BarrierLength"])
-
-    params["HdDiffu"] /= 10.0
-
-    params.pop("rmin")
-    params.pop("rmax")
-
-    # Maximum dune height
-    params["Dmaxel"] /= 10.0
-
-    # Erosion parameters
-    params["DuneRestart"] /= 10.0
-
-    # Volume of sediment lost via alongshore transport
-    params["Rat"] = params["Rat"] / -10.0
-    params["Qat"] = params["Rat"] * params["DShoreface"]
-    params.pop("Rat")
-
-    # Relative Sea Level Rise Rate
-    params["RSLR"] /= 10.0
-
-    params["surge_tide_m"] -= params["MHW"] * 10.0
-
-    params["Dshrub"] /= 10.0
-    params["ShrubEl_min"] = params["ShrubEl_min"] / 10.0 - params["MHW"]
-    params["ShrubEl_max"] = params["ShrubEl_max"] / 10.0 - params["MHW"]
-
-    # Overwash Interaction
-    params["BurialLimit"] /= 10.0
-    params["UprootLimit"] /= 10.0
-
-    # Percent cover change (years 0-9)
-    PC = np.array([0, 0.04, 0.08, 0.10, 0.15, 0.15, 0.20, 0.35, 0.80, 1])
-    addend = np.ones(params["TMAX"] + 50)  # (years 10+)
-    params["PC"] = np.append(PC, addend)
-
-    return params
